+++ lib/bugsnag/integrations/resque.rb
@@ -44,8 +44,18 @@ module Bugsnag
           :attributes => FRAMEWORK_ATTRIBUTES
         }
 
-        context = "#{payload['class']}@#{queue}"
-        report.meta_data.merge!({:context => context, :payload => payload})
+        metadata = payload
+        class_name = payload['class']
+
+        # when using Active Job the payload "class" will always be the Resque
+        # "JobWrapper", not the actual job class so we need to fix this here
+        if metadata['args'] && metadata['args'][0] && metadata['args'][0]['job_class']
+          class_name = metadata['args'][0]['job_class']
+          metadata['wrapped'] ||= class_name
+        end
+
+        context = "#{class_name}@#{queue}"
+        report.meta_data.merge!({ context: context, payload: metadata })
         report.context = context
       end
     end