+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Changelog
 * Add support for tracking exceptions and capturing metadata in Active Job, when not using an existing integration
   | [#670](https://github.com/bugsnag/bugsnag-ruby/pull/670)
 
+* Improve the report context when using Delayed Job or Resque as the Active Job queue adapter
+  | [#671](https://github.com/bugsnag/bugsnag-ruby/pull/671)
+
 ## v6.21.0 (23 June 2021)
 
 ### Enhancements